requirements.txt
fastapi
uvicorn
python-multipart
pydantic
redis
celery
diffusers
transformers
accelerate
torch
pillow
python-dotenv
boto3
